projects/tryon/cpvton/requirements.txt
opencv-python
easydict
pillow
numpy